Enum symrs::sym::Expr

pub enum Expr {
    Integer(BigInt),
    Sym(Symbol),
    Approx(f64),
    Neg(Box<Expr>),
    Sum(Vec<Expr>),
    Product(Vec<Expr>),
    Ratio(Box<Expr>, Box<Expr>),
    Pow(Box<Expr>, Box<Expr>),
    Undefined,
}

An owned expression representation.

Variants
Integer(BigInt)

Represent an integer.

Sym(Symbol)

Represent a symbol like x, y, or delta.

Approx(f64)

Represent a float approximate value.

Neg(Box<Expr>)

Represent negative of a expression.

Sum(Vec<Expr>)

Represent a sum of some expressions, like x + y + z.

Product(Vec<Expr>)

Represent a product of some expressions, like x * y * z.

Ratio(Box<Expr>, Box<Expr>)

Represent a ratio.

Pow(Box<Expr>, Box<Expr>)

Represent a power.

Undefined

Represent an undefined value.

Methods

impl Expr

Constructors.

pub fn integer<I: Into<BigInt>>(i: I) -> Expr

Construct a integer value.

pub fn symbol(s: &str) -> Expr

Construct a symbol.

pub fn approximate<F: Into<f64>>(f: F) -> Expr

Construct an approximate value.

pub fn negative<E: Into<Expr>>(e: E) -> Expr

Construct negative of an expression.

pub fn pow<E: Into<Expr>>(self, e: E) -> Expr

Construct a exponentiation expression.

impl Expr

Classify a Expr.

pub fn is_primitive(&self) -> bool

Check if the expression is a primitive.

Examples

use symrs::sym::{Expr, Symbol};

assert_eq!(Expr::integer(1).is_primitive(), true);
assert_eq!(Expr::approximate(1).is_primitive(), true);
let x = Symbol::new("x");
assert_eq!((x + 1).is_primitive(), false);

pub fn priority_rank(&self) -> i32

Get the priority rank when they are displayed. A lower number implys a higher priority. The value is intended to use in comparation context and may be changed later, so don't rely on a specific number.

The order of operators are as follows: (From higher to lower)

Scalars: Integer, Sym, Approx, Undefined.

Pow.

Neg.

Product, Ratio.

Sum.

Examples

use symrs::sym::{Expr, Symbol};

let x = Symbol::new("x");
assert!(Expr::integer(1).priority_rank() < (1i32 + x).priority_rank());